Added Room database for using app while device is offline

+++ app/src/main/java/uz/mrsolijon/weatherapp/data/remote/model/CurrentWeather.kt
@@ -4,7 +4,7 @@ data class CurrentWeather(
     val dt: Long,
     val temp: Float,
     val feels_like: Float,
-    val humidity: Int,
+    val humidity: Float,
     val wind_speed: Float,
     val weather: List<WeatherCondition>
 )
